bpf: tests: condense DSR IPIP test
Run the IPv4 and IPv6 test from a single file.
